+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Libraries are `Box`ed to avoid possible segmentation faults
 - Added an on exit function to `RunCallbacks`.
 - Terminal interface now has history that can be retrieved using Up and Down arrows
+- `:static-files` supports glob patterns.
 
 ## 0.16.0
 - Increase `libloading` dependency to `0.6`.